requirements.txt
asgiref==3.11.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.1
django-scheduler==0.10.2
djangorestframework==3.16.1
Flask==3.1.2
fonttools==4.61.1
icalendar==7.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
